* commit 'dd0ada27da29343e5cd50066801ae3e0794b8b6b':
  More compile-time tuning

+++ compiler/dex/portable/mir_to_gbc.cc
@@ -705,7 +705,7 @@ bool MirConverter::ConvertMIRNode(MIR* mir, BasicBlock* bb,
   /* Prep Src and Dest locations */
   int next_sreg = 0;
   int next_loc = 0;
-  int attrs = mir_graph_->oat_data_flow_attributes_[opcode];
+  uint64_t attrs = mir_graph_->oat_data_flow_attributes_[opcode];
   rl_src[0] = rl_src[1] = rl_src[2] = mir_graph_->GetBadLoc();
   if (attrs & DF_UA) {
     if (attrs & DF_A_WIDE) {
